Update linux recipe for OpenJDK 17.

+++ recipe/build.sh
@@ -16,53 +16,56 @@ fi
 function jdk_install
 {
   chmod +x bin/*
-  mkdir -p $INSTALL_DIR/bin
-  mv bin/* $INSTALL_DIR/bin/
-  ls -la $INSTALL_DIR/bin
+  mkdir -p $PREFIX/bin
+  mv bin/* $PREFIX/bin/
+  ls -la $PREFIX/bin
 
-  mkdir -p $INSTALL_DIR/include
-  mv include/* $INSTALL_DIR/include
+  mkdir -p $PREFIX/include
+  mv include/* $PREFIX/include
   if [ -e ./lib/jspawnhelper ]; then
     chmod +x ./lib/jspawnhelper
   fi
 
-  mkdir -p $INSTALL_DIR/lib
-  mv lib/* $INSTALL_DIR/lib
+  mkdir -p $PREFIX/lib
+  mv lib/* $PREFIX/lib
 
   if [ -f "DISCLAIMER" ]; then
-    mv DISCLAIMER $INSTALL_DIR/DISCLAIMER
+    mv DISCLAIMER $PREFIX/DISCLAIMER
   fi
 
-  mkdir -p $INSTALL_DIR/conf
-  mv conf/* $INSTALL_DIR/conf
+  mkdir -p $PREFIX/conf
+  mv conf/* $PREFIX/conf
 
-  mkdir -p $INSTALL_DIR/jmods
-  mv jmods/* $INSTALL_DIR/jmods
+  mkdir -p $PREFIX/jmods
+  mv jmods/* $PREFIX/jmods
 
-  mkdir -p $INSTALL_DIR/legal
-  mv legal/* $INSTALL_DIR/legal
+  mkdir -p $PREFIX/legal
+  mv legal/* $PREFIX/legal
 
-  mkdir -p $INSTALL_DIR/man/man1
-  mv man/man1/* $INSTALL_DIR/man/man1
+  mkdir -p $PREFIX/man/man1
+  mv man/man1/* $PREFIX/man/man1
   rm -rf man/man1
-  mv man/* $INSTALL_DIR/man
+  mv man/* $PREFIX/man
 }
 
-function source_build
-{
-  cd src
+chmod +x configure
 
-  chmod +x configure
+function double_build
+{
 
   if [[ "$target_platform" == linux* ]]; then
-    rm $PREFIX/include/iconv.h
+    if [[ -f $PREFIX/include/iconv.h ]]; then
+      rm $PREFIX/include/iconv.h
+    fi
   fi
 
 
   if [[ "${CONDA_BUILD_CROSS_COMPILATION:-0}" == 1 ]]; then
     (
       if [[ "$build_platform" == linux* ]]; then
-        rm $BUILD_PREFIX/include/iconv.h
+        if [[ -f $PREFIX/include/iconv.h ]]; then
+          rm $PREFIX/include/iconv.h
+        fi
       fi
 
       export CPATH=$BUILD_PREFIX/include
@@ -81,99 +84,103 @@ function source_build
       export -n CXXFLAGS
       unset CPPFLAGS
 
-      CFLAGS=$(echo $CFLAGS | sed 's/-mcpu=[a-z0-9]*//g' | sed 's/-mtune=[a-z0-9]*//g' | sed 's/-march=[a-z0-9]*//g')
-      CXXFLAGS=$(echo $CXXFLAGS | sed 's/-mcpu=[a-z0-9]*//g' | sed 's/-mtune=[a-z0-9]*//g' | sed 's/-march=[a-z0-9]*//g')
+      unset CFLAGS
+      unset CXXFLAGS
+      unset LDFLAGS
 
       ulimit -c unlimited
-      mkdir build-build
-        pushd build-build
+      mkdir source_build
+        pushd source_build
         ../configure \
-          --prefix=${BUILD_PREFIX} \
+          --prefix=${PREFIX} \
           --build=${BUILD} \
           --host=${BUILD} \
           --target=${BUILD} \
-          --with-extra-cflags="${CFLAGS//$PREFIX/$BUILD_PREFIX}" \
-          --with-extra-cxxflags="${CXXFLAGS//$PREFIX/$BUILD_PREFIX} -fpermissive" \
-          --with-extra-ldflags="${LDFLAGS//$PREFIX/$BUILD_PREFIX}" \
+          --with-extra-cflags="${CDA_CFLAGS}" \
+          --with-extra-cxxflags="${CDA_CXXFLAGS} -fpermissive" \
+          --with-extra-ldflags="${CDA_LDFLAGS} -lrt" \
           --with-stdc++lib=dynamic \
           --disable-warnings-as-errors \
-          --with-x=${BUILD_PREFIX} \
-          --with-cups=${BUILD_PREFIX} \
+          --with-x=${PREFIX} \
+          --with-cups=${PREFIX} \
           --with-freetype=system \
           --with-giflib=system \
           --with-libpng=system \
           --with-zlib=system \
-          --with-libjpeg=system \
           --with-lcms=system \
-          --with-fontconfig=${BUILD_PREFIX} \
-          --with-boot-jdk=$SRC_DIR/bootjdk
+          --with-fontconfig=${PREFIX} \
+          --with-boot-jdk=$SRC_DIR/bootstrap
+        make JOBS=$CPU_COUNT
         make JOBS=$CPU_COUNT images
       popd
     )
   fi
 
-
-  export CPATH=$PREFIX/include
-  export LIBRARY_PATH=$PREFIX/lib
-  export BUILD_CC=${CC_FOR_BUILD}
-  export BUILD_CXX=${CXX_FOR_BUILD}
-  export BUILD_CPP=${CXX_FOR_BUILD//+/p}
-  export BUILD_NM=$($CC_FOR_BUILD -print-prog-name=nm)
-  export BUILD_AR=$($CC_FOR_BUILD -print-prog-name=ar)
-  export BUILD_OBJCOPY=$($CC_FOR_BUILD -print-prog-name=objcopy)
-  export BUILD_STRIP=$($CC_FOR_BUILD -print-prog-name=strip)
-
-  export -n CFLAGS
-  export -n CXXFLAGS
-  export -n LDFLAGS
-
-  CONFIGURE_ARGS=""
-  if [[ "${CONDA_BUILD_CROSS_COMPILATION:-0}" == 1 ]]; then
-    CONFIGURE_ARGS="--with-build-jdk=$SRC_DIR/src/build-build/images/jdk"
-  fi
-
-  ./configure \
-    --prefix=$PREFIX \
-    --build=${BUILD} \
-    --host=${HOST} \
-    --target=${HOST} \
-    --with-extra-cflags="$CFLAGS" \
-    --with-extra-cxxflags="$CXXFLAGS -fpermissive" \
-    --with-extra-ldflags="$LDFLAGS" \
-    --with-x=$PREFIX \
-    --with-cups=$PREFIX \
-    --with-freetype=system \
-    --with-fontconfig=$PREFIX \
-    --with-giflib=system \
-    --with-libpng=system \
-    --with-zlib=system \
-    --with-libjpeg=system \
-    --with-lcms=system \
-    --with-stdc++lib=dynamic \
-    --disable-warnings-as-errors \
-    --with-boot-jdk=$SRC_DIR/bootjdk \
-    ${CONFIGURE_ARGS}
-
-  make JOBS=$CPU_COUNT
-  make JOBS=$CPU_COUNT images
 }
 
 if [[ "$target_platform" == linux* ]]; then 
-  export INSTALL_DIR=$SRC_DIR/bootjdk/
-  jdk_install
-  source_build
-  cd build/*/images/jdk
+  if [[ "${CONDA_BUILD_CROSS_COMPILATION:-0}" == 1 ]]; then
+    double_build
+  fi
 fi
 
-export INSTALL_DIR=$PREFIX
-jdk_install
+export BOOTSTRAP=$SRC_DIR/bootstrap
+export CFLAGS_FROM_CONDA=$(echo $CFLAGS | sed 's/-mcpu=[a-z0-9]*//g' | sed 's/-mtune=[a-z0-9]*//g' | sed 's/-march=[a-z0-9]*//g')
+export CXXFLAGS_FROM_CONDA=$(echo $CXXFLAGS | sed 's/-mcpu=[a-z0-9]*//g' | sed 's/-mtune=[a-z0-9]*//g' | sed 's/-march=[a-z0-9]*//g')
+export LDFLAGS_FROM_CONDA=$LDFLAGS
+export CPATH=$PREFIX/include
+export LIBRARY_PATH=$PREFIX/lib
+export BUILD_CC=${CC_FOR_BUILD}
+export BUILD_CXX=${CXX_FOR_BUILD}
+export BUILD_CPP=${CXX_FOR_BUILD//+/p}
+export BUILD_NM=$($CC_FOR_BUILD -print-prog-name=nm)
+export BUILD_AR=$($CC_FOR_BUILD -print-prog-name=ar)
+export BUILD_OBJCOPY=$($CC_FOR_BUILD -print-prog-name=objcopy)
+export BUILD_STRIP=$($CC_FOR_BUILD -print-prog-name=strip)
+
+export -n CFLAGS
+export -n CXXFLAGS
+export -n LDFLAGS
+export -n MAKEFLAGS
+
+CONFIGURE_ARGS=""
+if [[ "${CONDA_BUILD_CROSS_COMPILATION:-0}" == 1 ]]; then
+  CONFIGURE_ARGS="--with-build-jdk=$SRC_DIR/source_build/images/jdk"
+fi
+
+if [[ "${target_platform}" == osx* ]]; then
+  export BOOTSTRAP=$BOOTSTRAP/Contents/Home
+  export DSTROOT=$PREFIX
+fi
+
+./configure \
+  --prefix=$PREFIX \
+  --openjdk-target=$BUILD \
+  --with-extra-cflags="$CFLAGS_FROM_CONDA" \
+  --with-extra-cxxflags="$CXXFLAGS_FROM_CONDA -fpermissive" \
+  --with-extra-ldflags="$LDFLAGS_FROM_CONDA" \
+  --with-x=$SRC_DIR \
+  --with-cups=$PREFIX \
+  --with-freetype=bundled \
+  --with-fontconfig=$SRC_DIR \
+  --with-giflib=bundled \
+  --with-libpng=system \
+  --with-zlib=system \
+  --with-lcms=system \
+  --with-stdc++lib=dynamic \
+  --disable-warnings-as-errors \
+  --with-boot-jdk=$BOOTSTRAP \
+  ${CONFIGURE_ARGS}
+
+make JOBS=$CPU_COUNT
+make JOBS=$CPU_COUNT images
+make install
 
 if [[ "$target_platform" == linux* ]]; then
-  mv $INSTALL_DIR/lib/jli/*.so $INSTALL_DIR/lib/
   # Include dejavu fonts to allow java to work even on minimal cloud
   # images where these fonts are missing (thanks to @chapmanb)
-  mkdir -p $INSTALL_DIR/lib/fonts
-  mv $SRC_DIR/fonts/ttf/* $INSTALL_DIR/lib/fonts/
+  mkdir -p $PREFIX/lib/fonts
+  mv $SRC_DIR/fonts/ttf/* $PREFIX/lib/fonts/
 fi
 find $PREFIX -name "*.debuginfo" -exec rm -rf {} \;
